removed statusCodes const, moved auth controller to its own folder
