+++ controllers/auth/auth.js
@@ -6,14 +6,7 @@ const {
     ALREADY_REGISTERED,
     PASS_LENGTH,
     INCORRECT_CREDENTIALS
-  }, 
-  statusCodes: {
-    UNAUTHORIZED,
-    CONFLICT,
-    UNPROCESSABLE,
-    OK,
-    CREATED,
-  } 
+  }
 } = require('~/consts/errors')
 
 const signup = async (req, res, next) => {
@@ -21,14 +14,14 @@ const signup = async (req, res, next) => {
 
   try {
     const candidate = await User.findOne({ email })
-    if (candidate) throw new ApiError (CONFLICT, ALREADY_REGISTERED)
+    if (candidate) throw new ApiError (409, ALREADY_REGISTERED)
 
-    if (password.length < 8 || password.length > 25) throw new ApiError (UNPROCESSABLE, PASS_LENGTH)
+    if (password.length < 8 || password.length > 25) throw new ApiError (422, PASS_LENGTH)
     
     const hashedPassword = await hashPassword(password)
     const user = await User.create({ role, firstName, lastName, email, password: hashedPassword })
 
-    res.status(CREATED).json({ user: { firstName, lastName, email, id: user._id } })
+    res.status(201).json({ user: { firstName, lastName, email, id: user._id } })
   } catch (err) {
     next(err)
   }
@@ -39,14 +32,14 @@ const login = async (req, res, next) => {
 
   try {
     const user = await User.findOne({ email })
-    if (!user) throw new ApiError (UNAUTHORIZED, INCORRECT_CREDENTIALS)
+    if (!user) throw new ApiError (401, INCORRECT_CREDENTIALS)
 
     const auth = await comparePasswords(password, user.password)
-    if (!auth) throw new ApiError (UNAUTHORIZED, INCORRECT_CREDENTIALS)
+    if (!auth) throw new ApiError (401, INCORRECT_CREDENTIALS)
 
     const token = createToken(user._id)
 
-    res.status(OK).json({ user: { firstName: user.firstName, lastName: user.lastName, email, id: user._id } })
+    res.status(200).json({ user: { firstName: user.firstName, lastName: user.lastName, email, id: user._id } })
   } catch (err) {
     next(err)
   }
