+++ routes/auth.js
@@ -1,6 +1,6 @@
 const express = require('express')
 
-const auth = require('~/controllers/auth')
+const auth = require('~/controllers/auth/auth')
 
 const router = express.Router()
